refactor(Article): move to src/pages
